refactor(vis): move all code to single package
