--- packages/visualizer-core/README.md
@@ -1,1 +1,0 @@
-# visualizer-core
