--- packages/visualizer-core/moon.yml
@@ -1,2 +1,0 @@
-language: typescript
-type: library
