--- packages/visualizer-core/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "@moonrepo/visualizer-core",
-  "devDependencies": {
-    "@types/cytoscape": "3.19.9",
-    "packemon": "^2.4.1"
-  },
-  "packemon": {
-    "format": "cjs",
-    "platform": [
-      "browser",
-      "node"
-    ],
-    "bundle": true
-  },
-  "types": "./dts/index.d.ts",
-  "main": "./cjs/index.cjs",
-  "module": "./esm/index.js",
-  "exports": {
-    "./package.json": "./package.json",
-    ".": {
-      "types": "./dts/index.d.ts",
-      "browser": {
-        "module": "./esm/index.js",
-        "import": "./esm/index.js",
-        "default": "./lib/index.js"
-      },
-      "node": {
-        "import": "./cjs/index-wrapper.mjs",
-        "require": "./cjs/index.cjs"
-      },
-      "default": "./lib/index.js"
-    }
-  },
-  "files": [
-    "cjs/**/*.{cjs,mjs,map}",
-    "dts/**/*.d.ts",
-    "esm/**/*.{js,map}",
-    "lib/**/*.{js,map}",
-    "src/**/*.{ts,tsx,json}"
-  ],
-  "dependencies": {
-    "cytoscape": "3.23.0"
-  }
-}
