--- packages/visualizer-core/tsconfig.build.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dts",
-    "rootDir": "src"
-  },
-  "include": [
-    "src/**/*"
-  ]
-}
