--- packages/visualizer-core/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.options.json",
-  "include": [
-    "**/*"
-  ],
-  "references": [],
-  "compilerOptions": {
-    "outDir": "../../.moon/cache/types/packages/visualizer-core"
-  }
-}
