+++ packages/visualizer/package.json
@@ -4,11 +4,12 @@
   "version": "0.0.0",
   "type": "module",
   "dependencies": {
-    "@moonrepo/visualizer-core": "workspace:*",
+    "cytoscape": "3.23.0",
     "preact": "^10.11.2"
   },
   "devDependencies": {
     "@preact/preset-vite": "^2.4.0",
+    "@types/cytoscape": "3.19.9",
     "eslint-config-preact": "^1.3.0",
     "vite": "^3.2.3",
     "vite-plugin-css-injected-by-js": "^2.1.1"
