+++ packages/visualizer/src/lib/render.ts
@@ -1,7 +1,5 @@
-import cytoscape from 'cytoscape';
-import { GraphInfo } from './types';
-
-export * from './types';
+import cytoscape from "cytoscape";
+import { GraphInfo } from "./types";
 
 export const render = (element: HTMLElement, data: GraphInfo) => {
 	const nodes = data.nodes.map((n) => ({
